DRAW-351 refactor: Session 객체 WebSocket으로 이관
