+++ domain/src/main/kotlin/com/xorker/draw/websocket/Session.kt
@@ -1,17 +1,9 @@
 package com.xorker.draw.websocket
 
-import com.xorker.draw.room.RoomId
 import com.xorker.draw.user.User
 
 @JvmInline
 value class SessionId(val value: String)
-
-interface Session {
-    val id: SessionId
-    val user: User
-    val roomId: RoomId
-    fun send(message: String)
-}
 
 interface WaitingQueueSession {
     val id: SessionId
